solve the issue with topic auth and some minor errors

+++ telegram_bot/routers/topic_bot/topic_delete.py
@@ -6,12 +6,35 @@ from .topic_router import router
 import httpx
 
 API_URL = "http://127.0.0.1:8000/topic/{topic_id}"
+API_GET_URL = "http://127.0.0.1:8000/token/{telegram_id}"
+API_TOKEN_URL = "http://127.0.0.1:8000/auth/validate"
 
 class DeleteTopicState(StatesGroup):
     waiting_id = State()
 
 @router.message(Command("delete_topic"))
 async def start_get(message: types.Message, state: FSMContext):
+    telegram_id = message.from_user.id
+    async with httpx.AsyncClient() as client:
+        response = await client.get(API_GET_URL.format(telegram_id=telegram_id))
+
+    if response.status_code != 200:
+        await message.answer(f"User with telegram id {telegram_id} unauthorized. Use command /login for authorize")
+        await state.clear()
+        return
+
+    token = response.json().get("access_token")
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get(API_TOKEN_URL, headers={"Authorization": f"Bearer {token}"})
+
+    if response.status_code != 200:
+        await message.answer(f"User didn't authorize. Use command /login for authorize")
+        await state.clear()
+        return
+
+    await state.update_data(token=token)
+
     await message.answer("Enter the id of topic:")
     await state.set_state(DeleteTopicState.waiting_id)
 
@@ -25,8 +48,10 @@ async def get_topic(message: types.Message, state: FSMContext):
         await message.answer("Enter a integer number")
         return
 
+    data = await state.get_data()
+    token = data.pop("token")
     async with httpx.AsyncClient() as client:
-        response = await client.delete(API_URL.format(topic_id=topic_id))
+        response = await client.delete(API_URL.format(topic_id=topic_id), headers={"Authorization": f"Bearer {token}"})
 
     if response.status_code == 200:
         await message.answer("Topic deleted successfully ✅")
